smart_hotel_backend/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
pandas
python-dotenv